Preserve accuracy routing and expert accumulation semantics
Keep FP32 accuracy routing, shared expert ordering and native expert accumulation behavior consistent with the validated candidates.

+++ megatron/core/transformer/moe/moe_utils.py
@@ -880,13 +880,13 @@ def topk_routing_with_score_function(
     elif score_function in ("sigmoid", "sqrtsoftplus"):
         if _use_accuracy_compatible():
             if score_function == "sigmoid":
-                scores = torch.sigmoid(logits.float()).type_as(logits)
+                scores = torch.sigmoid(logits.float())
             else:
-                scores = torch.nn.functional.softplus(logits.float()).sqrt().type_as(logits)
+                scores = torch.nn.functional.softplus(logits.float()).sqrt()
             if expert_bias is not None:
                 scores_for_routing = scores + expert_bias
                 _, top_indices = compute_topk(scores_for_routing, topk, num_groups, group_topk)
-                scores = torch.gather(scores, dim=1, index=top_indices).type_as(logits)
+                scores = torch.gather(scores, dim=1, index=top_indices)
             else:
                 scores, top_indices = compute_topk(scores, topk, num_groups, group_topk)
             _scores_f64 = scores.double()
